Wishlist method
Finally its perfectly working as expected with no errors. there are two approaches to implement it. The second one works betters and is not just limited to one device because its cloud based. It reads, retrieves and removes data from the cloud. whereas the first approach was device based and it wouldn't have worked perfectly once a different device